requirements.txt
librosa>=0.9.2
soundfile>=0.12.1
numpy>=1.23
torch>=2.0
torchaudio>=2.0
torchcodec>=0.8.0
tqdm>=4.64
coqui-tts
transformers>=4.57,<5
scipy>=1.9
requests==2.32.4
beautifulsoup4>=4.11
pandas<3.0
